+++ CHANGELOG.md
@@ -5,7 +5,10 @@
 ### Added
 
 - concord-server, concord-console: support for `yield` for
-non-custom forms.
+non-custom forms;
+- ansible: support for external private key files;
+- ansible: support for external configuration files;
+- ansible: verbosity level can now be set using the task's arguments.
 
 ### Changed